+++ NEWS.md
@@ -1,3 +1,15 @@
+# rangl dev
+
+* raster package is now an Import
+
+* added support for RasterLayer
+
+* fixed globe to keep PROJ.4
+
+* quashed a major bug introduced by use of dplyr::distinct, best to use factor unique classifier on character versions of coords
+
+* several cleanup fixes
+
 # rangl 0.3.0
 
 * rename again, main function is called 'rangl', the term 'mesh' is too often used across R