+++ NEWS.md
@@ -1,6 +1,7 @@
 # rangl dev
 
 * points now have meta table (it was missing), and singular points are now supported
+
 * raster package is now an Import
 
 * added support for RasterLayer