+++ NEWS.md
@@ -1,5 +1,4 @@
-# rangl dev
-
+# rangl 0.4.0
 * points now have meta table (it was missing), and singular points are now supported
 
 * raster package is now an Import